# LMS/docs/swagger.yaml
basePath: /lmsapiv1
definitions:
...
  dto.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  dto.UserRolesResponse:
    properties:
      email:
        type: string
      roles:
        items:
          type: string
        type: array
      user_id:
        type: integer
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@bdc.hpcc.vn
    name: API Support
    url: https://bdc.hpcc.vn/support
  description: This is the API documentation for LMS (Learning Management System)
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://bdc.hpcc.vn/terms
  title: LMS API Documentation
  version: "1.0"
paths:
  /attempts/{attemptId}/summary:
    get:
      consumes:
      - application/json
      description: Get detailed summary of a quiz attempt including statistics
      parameters:
      - description: Attempt ID
        in: path
        name: attemptId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Attempt summary
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid attempt ID
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Not authorized to view this attempt
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Attempt not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get summary of a specific attempt
      tags:
      - Quiz - Student
...
  /me/roles:
    get:
      consumes:
      - application/json
      description: Get the roles of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserRolesResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get my roles
      tags:
      - users
...
  /quizzes/{quizId}/my-attempts:
    get:
      consumes:
      - application/json
      description: Get all attempts by the current student for a specific quiz
      parameters:
      - description: Quiz ID
        in: path
        name: quizId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of attempts
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid quiz ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get student's quiz attempt history
      tags:
      - Quiz - Student
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"